Move main logic to  package and use generics for API with backwards compatibility
