+++ go.mod
@@ -1,5 +1,10 @@
 module github.com/jackc/puddle
 
-go 1.12
+go 1.18
 
 require github.com/stretchr/testify v1.3.0
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+)
