+++ puddleg/nanotime_unsafe.go
@@ -1,8 +1,8 @@
-// +build !purego,!appengine,!js
+//go:build !purego && !appengine && !js
 
 // This file contains the implementation of nanotime using runtime.nanotime.
 
-package puddle
+package puddleg
 
 import "unsafe"
