+++ puddleg/pool_test.go
@@ -1,4 +1,4 @@
-package puddle_test
+package puddleg_test
 
 import (
 	"context"
@@ -14,7 +14,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/jackc/puddle"
+	"github.com/jackc/puddle/puddleg"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -41,48 +41,25 @@ func (c *Counter) Value() int {
 	return n
 }
 
-func createConstructor() (puddle.Constructor, *Counter) {
+func createConstructor() (puddleg.Constructor[int], *Counter) {
 	var c Counter
-	f := func(ctx context.Context) (interface{}, error) {
+	f := func(ctx context.Context) (int, error) {
 		return c.Next(), nil
 	}
 	return f, &c
 }
 
-func createConstructorWithNotifierChan() (puddle.Constructor, *Counter, chan int) {
-	ch := make(chan int)
-	var c Counter
-	f := func(ctx context.Context) (interface{}, error) {
-		n := c.Next()
-
-		// Because the tests will not read from ch until after the create function f returns.
-		go func() { ch <- n }()
-
-		return n, nil
-	}
-	return f, &c, ch
-}
-
-func stubDestructor(interface{}) {}
-
-func waitForRead(ch chan int) bool {
-	select {
-	case <-ch:
-		return true
-	case <-time.NewTimer(time.Second).C:
-		return false
-	}
-}
+func stubDestructor(int) {}
 
 func TestNewPoolRequiresMaxSizeGreaterThan0(t *testing.T) {
 	constructor, _ := createConstructor()
-	assert.Panics(t, func() { puddle.NewPool(constructor, stubDestructor, -1) })
-	assert.Panics(t, func() { puddle.NewPool(constructor, stubDestructor, 0) })
+	assert.Panics(t, func() { puddleg.NewPool(constructor, stubDestructor, -1) })
+	assert.Panics(t, func() { puddleg.NewPool(constructor, stubDestructor, 0) })
 }
 
 func TestPoolAcquireCreatesResourceWhenNoneIdle(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 	defer pool.Close()
 
 	res, err := pool.Acquire(context.Background())
@@ -94,7 +71,7 @@ func TestPoolAcquireCreatesResourceWhenNoneIdle(t *testing.T) {
 
 func TestPoolAcquireDoesNotCreatesResourceWhenItWouldExceedMaxSize(t *testing.T) {
 	constructor, createCounter := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 1)
+	pool := puddleg.NewPool(constructor, stubDestructor, 1)
 
 	wg := &sync.WaitGroup{}
 
@@ -119,7 +96,7 @@ func TestPoolAcquireDoesNotCreatesResourceWhenItWouldExceedMaxSize(t *testing.T)
 
 func TestPoolAcquireWithCancellableContext(t *testing.T) {
 	constructor, createCounter := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 1)
+	pool := puddleg.NewPool(constructor, stubDestructor, 1)
 
 	wg := &sync.WaitGroup{}
 
@@ -146,10 +123,10 @@ func TestPoolAcquireWithCancellableContext(t *testing.T) {
 
 func TestPoolAcquireReturnsErrorFromFailedResourceCreate(t *testing.T) {
 	errCreateFailed := errors.New("create failed")
-	constructor := func(ctx context.Context) (interface{}, error) {
-		return nil, errCreateFailed
+	constructor := func(ctx context.Context) (int, error) {
+		return 0, errCreateFailed
 	}
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	res, err := pool.Acquire(context.Background())
 	assert.Equal(t, errCreateFailed, err)
@@ -158,7 +135,7 @@ func TestPoolAcquireReturnsErrorFromFailedResourceCreate(t *testing.T) {
 
 func TestPoolAcquireReusesResources(t *testing.T) {
 	constructor, createCounter := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	res, err := pool.Acquire(context.Background())
 	require.NoError(t, err)
@@ -177,11 +154,11 @@ func TestPoolAcquireReusesResources(t *testing.T) {
 
 func TestPoolTryAcquire(t *testing.T) {
 	constructor, createCounter := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 1)
+	pool := puddleg.NewPool(constructor, stubDestructor, 1)
 
 	// Pool is initially empty so TryAcquire fails but starts construction of resource in the background.
 	res, err := pool.TryAcquire(context.Background())
-	require.EqualError(t, err, puddle.ErrNotAvailable.Error())
+	require.EqualError(t, err, puddleg.ErrNotAvailable.Error())
 	assert.Nil(t, res)
 
 	// Wait for background creation to complete.
@@ -193,7 +170,7 @@ func TestPoolTryAcquire(t *testing.T) {
 	defer res.Release()
 
 	res, err = pool.TryAcquire(context.Background())
-	require.EqualError(t, err, puddle.ErrNotAvailable.Error())
+	require.EqualError(t, err, puddleg.ErrNotAvailable.Error())
 	assert.Nil(t, res)
 
 	assert.Equal(t, 1, createCounter.Value())
@@ -201,29 +178,29 @@ func TestPoolTryAcquire(t *testing.T) {
 
 func TestPoolTryAcquireReturnsErrorWhenPoolIsClosed(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 	pool.Close()
 
 	res, err := pool.TryAcquire(context.Background())
-	assert.Equal(t, puddle.ErrClosedPool, err)
+	assert.Equal(t, puddleg.ErrClosedPool, err)
 	assert.Nil(t, res)
 }
 
 func TestPoolTryAcquireWithFailedResourceCreate(t *testing.T) {
 	errCreateFailed := errors.New("create failed")
-	constructor := func(ctx context.Context) (interface{}, error) {
-		return nil, errCreateFailed
+	constructor := func(ctx context.Context) (int, error) {
+		return 0, errCreateFailed
 	}
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	res, err := pool.TryAcquire(context.Background())
-	require.EqualError(t, err, puddle.ErrNotAvailable.Error())
+	require.EqualError(t, err, puddleg.ErrNotAvailable.Error())
 	assert.Nil(t, res)
 }
 
 func TestPoolAcquireNilContextDoesNotLeavePoolLocked(t *testing.T) {
 	constructor, createCounter := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	assert.Panics(t, func() { pool.Acquire(nil) })
 
@@ -236,10 +213,10 @@ func TestPoolAcquireNilContextDoesNotLeavePoolLocked(t *testing.T) {
 }
 
 func TestPoolAcquireContextAlreadyCanceled(t *testing.T) {
-	constructor := func(ctx context.Context) (interface{}, error) {
+	constructor := func(ctx context.Context) (int, error) {
 		panic("should never be called")
 	}
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	ctx, cancel := context.WithCancel(context.Background())
 	cancel()
@@ -254,15 +231,15 @@ func TestPoolAcquireContextCanceledDuringCreate(t *testing.T) {
 	timeoutChan := time.After(1 * time.Second)
 
 	var constructorCalls Counter
-	constructor := func(ctx context.Context) (interface{}, error) {
+	constructor := func(ctx context.Context) (int, error) {
 		select {
 		case <-ctx.Done():
-			return nil, ctx.Err()
+			return 0, ctx.Err()
 		case <-timeoutChan:
 		}
 		return constructorCalls.Next(), nil
 	}
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	res, err := pool.Acquire(ctx)
 	assert.Equal(t, context.Canceled, err)
@@ -271,10 +248,10 @@ func TestPoolAcquireContextCanceledDuringCreate(t *testing.T) {
 
 func TestPoolAcquireAllIdle(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 	defer pool.Close()
 
-	resources := make([]*puddle.Resource, 4)
+	resources := make([]*puddleg.Resource[int], 4)
 	var err error
 
 	resources[0], err = pool.Acquire(context.Background())
@@ -291,7 +268,7 @@ func TestPoolAcquireAllIdle(t *testing.T) {
 	resources[0].Release()
 	resources[3].Release()
 
-	assert.ElementsMatch(t, []*puddle.Resource{resources[0], resources[3]}, pool.AcquireAllIdle())
+	assert.ElementsMatch(t, []*puddleg.Resource[int]{resources[0], resources[3]}, pool.AcquireAllIdle())
 
 	resources[0].Release()
 	resources[3].Release()
@@ -308,14 +285,14 @@ func TestPoolAcquireAllIdle(t *testing.T) {
 
 func TestPoolAcquireAllIdleWhenClosedIsNil(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 	pool.Close()
 	assert.Nil(t, pool.AcquireAllIdle())
 }
 
 func TestPoolCreateResource(t *testing.T) {
 	constructor, counter := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 	defer pool.Close()
 
 	var err error
@@ -339,10 +316,10 @@ func TestPoolCreateResource(t *testing.T) {
 
 func TestPoolCreateResourceReturnsErrorFromFailedResourceCreate(t *testing.T) {
 	errCreateFailed := errors.New("create failed")
-	constructor := func(ctx context.Context) (interface{}, error) {
-		return nil, errCreateFailed
+	constructor := func(ctx context.Context) (int, error) {
+		return 0, errCreateFailed
 	}
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	err := pool.CreateResource(context.Background())
 	assert.Equal(t, errCreateFailed, err)
@@ -350,20 +327,20 @@ func TestPoolCreateResourceReturnsErrorFromFailedResourceCreate(t *testing.T) {
 
 func TestPoolCreateResourceReturnsErrorWhenAlreadyClosed(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 	pool.Close()
 	err := pool.CreateResource(context.Background())
-	assert.Equal(t, puddle.ErrClosedPool, err)
+	assert.Equal(t, puddleg.ErrClosedPool, err)
 }
 
 func TestPoolCreateResourceReturnsErrorWhenClosedWhileCreatingResource(t *testing.T) {
 	// There is no way to guarantee the correct order of the pool being closed while the resource is being constructed.
 	// But these sleeps should make it extremely likely. (Ah, the lengths we go for 100% test coverage...)
-	constructor := func(ctx context.Context) (interface{}, error) {
+	constructor := func(ctx context.Context) (int, error) {
 		time.Sleep(500 * time.Millisecond)
-		return "abc", nil
+		return 123, nil
 	}
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	acquireErrChan := make(chan error)
 	go func() {
@@ -375,20 +352,20 @@ func TestPoolCreateResourceReturnsErrorWhenClosedWhileCreatingResource(t *testin
 	pool.Close()
 
 	err := <-acquireErrChan
-	assert.Equal(t, puddle.ErrClosedPool, err)
+	assert.Equal(t, puddleg.ErrClosedPool, err)
 }
 
 func TestPoolCloseClosesAllIdleResources(t *testing.T) {
 	constructor, _ := createConstructor()
 
 	var destructorCalls Counter
-	destructor := func(interface{}) {
+	destructor := func(int) {
 		destructorCalls.Next()
 	}
 
-	p := puddle.NewPool(constructor, destructor, 10)
+	p := puddleg.NewPool(constructor, destructor, 10)
 
-	resources := make([]*puddle.Resource, 4)
+	resources := make([]*puddleg.Resource[int], 4)
 	for i := range resources {
 		var err error
 		resources[i], err = p.Acquire(context.Background())
@@ -407,13 +384,13 @@ func TestPoolCloseClosesAllIdleResources(t *testing.T) {
 func TestPoolCloseBlocksUntilAllResourcesReleasedAndClosed(t *testing.T) {
 	constructor, _ := createConstructor()
 	var destructorCalls Counter
-	destructor := func(interface{}) {
+	destructor := func(int) {
 		destructorCalls.Next()
 	}
 
-	p := puddle.NewPool(constructor, destructor, 10)
+	p := puddleg.NewPool(constructor, destructor, 10)
 
-	resources := make([]*puddle.Resource, 4)
+	resources := make([]*puddleg.Resource[int], 4)
 	for i := range resources {
 		var err error
 		resources[i], err = p.Acquire(context.Background())
@@ -421,7 +398,7 @@ func TestPoolCloseBlocksUntilAllResourcesReleasedAndClosed(t *testing.T) {
 	}
 
 	for _, res := range resources {
-		go func(res *puddle.Resource) {
+		go func(res *puddleg.Resource[int]) {
 			time.Sleep(100 * time.Millisecond)
 			res.Release()
 		}(res)
@@ -434,7 +411,7 @@ func TestPoolCloseBlocksUntilAllResourcesReleasedAndClosed(t *testing.T) {
 func TestPoolCloseIsSafeToCallMultipleTimes(t *testing.T) {
 	constructor, _ := createConstructor()
 
-	p := puddle.NewPool(constructor, stubDestructor, 10)
+	p := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	p.Close()
 	p.Close()
@@ -446,7 +423,7 @@ func TestPoolStatResources(t *testing.T) {
 	endWaitChan := make(chan struct{})
 
 	var constructorCalls Counter
-	constructor := func(ctx context.Context) (interface{}, error) {
+	constructor := func(ctx context.Context) (int, error) {
 		select {
 		case <-startWaitChan:
 			close(waitingChan)
@@ -456,7 +433,7 @@ func TestPoolStatResources(t *testing.T) {
 
 		return constructorCalls.Next(), nil
 	}
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 	defer pool.Close()
 
 	resAcquired, err := pool.Acquire(context.Background())
@@ -491,7 +468,7 @@ func TestPoolStatResources(t *testing.T) {
 
 func TestPoolStatSuccessfulAcquireCounters(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 1)
+	pool := puddleg.NewPool(constructor, stubDestructor, 1)
 	defer pool.Close()
 
 	res, err := pool.Acquire(context.Background())
@@ -537,7 +514,7 @@ func TestPoolStatSuccessfulAcquireCounters(t *testing.T) {
 
 func TestPoolStatCanceledAcquireBeforeStart(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 1)
+	pool := puddleg.NewPool(constructor, stubDestructor, 1)
 	defer pool.Close()
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -551,12 +528,12 @@ func TestPoolStatCanceledAcquireBeforeStart(t *testing.T) {
 }
 
 func TestPoolStatCanceledAcquireDuringCreate(t *testing.T) {
-	constructor := func(ctx context.Context) (interface{}, error) {
+	constructor := func(ctx context.Context) (int, error) {
 		<-ctx.Done()
-		return nil, ctx.Err()
+		return 0, ctx.Err()
 	}
 
-	pool := puddle.NewPool(constructor, stubDestructor, 1)
+	pool := puddleg.NewPool(constructor, stubDestructor, 1)
 	defer pool.Close()
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -571,7 +548,7 @@ func TestPoolStatCanceledAcquireDuringCreate(t *testing.T) {
 
 func TestPoolStatCanceledAcquireDuringWait(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 1)
+	pool := puddleg.NewPool(constructor, stubDestructor, 1)
 	defer pool.Close()
 
 	res, err := pool.Acquire(context.Background())
@@ -592,11 +569,11 @@ func TestPoolStatCanceledAcquireDuringWait(t *testing.T) {
 func TestResourceHijackRemovesResourceFromPoolButDoesNotDestroy(t *testing.T) {
 	constructor, _ := createConstructor()
 	var destructorCalls Counter
-	destructor := func(interface{}) {
+	destructor := func(int) {
 		destructorCalls.Next()
 	}
 
-	pool := puddle.NewPool(constructor, destructor, 10)
+	pool := puddleg.NewPool(constructor, destructor, 10)
 
 	res, err := pool.Acquire(context.Background())
 	require.NoError(t, err)
@@ -615,7 +592,7 @@ func TestResourceHijackRemovesResourceFromPoolButDoesNotDestroy(t *testing.T) {
 
 func TestResourceDestroyRemovesResourceFromPool(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	res, err := pool.Acquire(context.Background())
 	require.NoError(t, err)
@@ -635,7 +612,7 @@ func TestResourceDestroyRemovesResourceFromPool(t *testing.T) {
 
 func TestResourceLastUsageTimeTracking(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 1)
+	pool := puddleg.NewPool(constructor, stubDestructor, 1)
 
 	res, err := pool.Acquire(context.Background())
 	require.NoError(t, err)
@@ -669,7 +646,7 @@ func TestResourceLastUsageTimeTracking(t *testing.T) {
 
 func TestResourcePanicsOnUsageWhenNotAcquired(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 
 	res, err := pool.Acquire(context.Background())
 	require.NoError(t, err)
@@ -687,11 +664,11 @@ func TestResourcePanicsOnUsageWhenNotAcquired(t *testing.T) {
 
 func TestPoolAcquireReturnsErrorWhenPoolIsClosed(t *testing.T) {
 	constructor, _ := createConstructor()
-	pool := puddle.NewPool(constructor, stubDestructor, 10)
+	pool := puddleg.NewPool(constructor, stubDestructor, 10)
 	pool.Close()
 
 	res, err := pool.Acquire(context.Background())
-	assert.Equal(t, puddle.ErrClosedPool, err)
+	assert.Equal(t, puddleg.ErrClosedPool, err)
 	assert.Nil(t, res)
 }
 
@@ -705,7 +682,7 @@ func TestSignalIsSentWhenResourceFailedToCreate(t *testing.T) {
 	}
 	destructor := func(value interface{}) {}
 
-	pool := puddle.NewPool(constructor, destructor, 1)
+	pool := puddleg.NewPool(constructor, destructor, 1)
 
 	res1, err := pool.Acquire(context.Background())
 	require.NoError(t, err)
@@ -728,7 +705,7 @@ func TestSignalIsSentWhenResourceFailedToCreate(t *testing.T) {
 func TestStress(t *testing.T) {
 	constructor, _ := createConstructor()
 	var destructorCalls Counter
-	destructor := func(interface{}) {
+	destructor := func(int) {
 		destructorCalls.Next()
 	}
 
@@ -737,16 +714,16 @@ func TestStress(t *testing.T) {
 		poolSize = 4
 	}
 
-	pool := puddle.NewPool(constructor, destructor, int32(poolSize))
+	pool := puddleg.NewPool(constructor, destructor, int32(poolSize))
 
 	finishChan := make(chan struct{})
 	wg := &sync.WaitGroup{}
 
-	releaseOrDestroyOrHijack := func(res *puddle.Resource) {
+	releaseOrDestroyOrHijack := func(res *puddleg.Resource[int]) {
 		n := rand.Intn(100)
 		if n < 5 {
 			res.Hijack()
-			destructor(res)
+			destructor(res.Value())
 		} else if n < 10 {
 			res.Destroy()
 		} else {
@@ -759,7 +736,7 @@ func TestStress(t *testing.T) {
 		func() {
 			res, err := pool.Acquire(context.Background())
 			if err != nil {
-				if err != puddle.ErrClosedPool {
+				if err != puddleg.ErrClosedPool {
 					assert.Failf(t, "stress acquire", "pool.Acquire returned unexpected err: %v", err)
 				}
 				return
@@ -774,7 +751,7 @@ func TestStress(t *testing.T) {
 			defer cancel()
 			res, err := pool.Acquire(ctx)
 			if err != nil {
-				if err != puddle.ErrClosedPool && err != context.Canceled && err != context.DeadlineExceeded {
+				if err != puddleg.ErrClosedPool && err != context.Canceled && err != context.DeadlineExceeded {
 					assert.Failf(t, "stress acquire possibly canceled by context", "pool.Acquire returned unexpected err: %v", err)
 				}
 				return
@@ -787,7 +764,7 @@ func TestStress(t *testing.T) {
 		func() {
 			res, err := pool.TryAcquire(context.Background())
 			if err != nil {
-				if err != puddle.ErrClosedPool && err != puddle.ErrNotAvailable {
+				if err != puddleg.ErrClosedPool && err != puddleg.ErrNotAvailable {
 					assert.Failf(t, "stress TryAcquire", "pool.TryAcquire returned unexpected err: %v", err)
 				}
 				return
@@ -873,7 +850,7 @@ func ExamplePool() {
 	}
 	maxPoolSize := int32(10)
 
-	pool := puddle.NewPool(constructor, destructor, maxPoolSize)
+	pool := puddleg.NewPool(constructor, destructor, maxPoolSize)
 
 	// Use pool multiple times
 	for i := 0; i < 10; i++ {
@@ -972,7 +949,7 @@ func BenchmarkPoolAcquireAndRelease(b *testing.B) {
 			wg := &sync.WaitGroup{}
 
 			constructor, _ := createConstructor()
-			pool := puddle.NewPool(constructor, stubDestructor, bm.poolSize)
+			pool := puddleg.NewPool(constructor, stubDestructor, bm.poolSize)
 
 			for i := 0; i < bm.clientCount; i++ {
 				wg.Add(1)
